Camada Controls: Adicionado as classes LoginMethods e ChangePasswordMethods, para que as mesmas recebam os métodos de seus respectivos forms
