+++ QrAguas/QrAguas.csproj
@@ -94,6 +94,8 @@
   <ItemGroup>
     <Compile Include="BusinessLayer\BlDataVerification.cs" />
     <Compile Include="Controls\Connection.cs" />
+    <Compile Include="Controls\ChangePasswordMethods.cs" />
+    <Compile Include="Controls\LoginMethods.cs" />
     <Compile Include="Controls\Functions.cs" />
     <Compile Include="Models\CepResponse.cs" />
     <Compile Include="Models\ICepApiService.cs" />
